+++ src/signatures.rs
@@ -259,6 +259,9 @@ signatures! {
     value = b"\xEF\xBB\xBF<USFSubtitles"
     value = b"<USFSubtitles"
 
+    format = WindowsRecordedTvShow
+    value = b"\xB7\xD8\x00\x20\x37\x49\xDA\x11\xA6\x4E\x00\x07\xE9\x5E\xAD\x8D"
+
     // 15 bytes
     format = Fictionbook
     value = b"\xEF\xBB\xBF<FictionBook"